Full_Stack_WebApp/Backend_WebApp/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.3
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
flake8==7.0.0
Pillow 